Fixed formatting of Gantt content.
